[TEST: REFACTOR] Rename MockUserRepository
